requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
configparser==5.0.0
Django==3.0.7
django-rest-registration==0.5.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
typing==3.7.4.1
urllib3==1.25.9
